Added an Options Menu that is in the same scene as the Main Menu. Added functionality of adjusting volumes and saving settings.
Options Menu has a volume slider which temporarily adjusts the volume of the player's current game session. Should they choose to Apply the settings, then the settings are applied to the current game session, and saved to the player preferences file. Should they choose to Exit the options menu, then the settings will be reverted to what they were before the Player had applied them.